+++ src/api/api.ts
@@ -1,12 +1,4 @@
-import axios from 'axios';
-
-const BASE_URL = 'https://pokeapi.co/api/v2/';
-
-const instance = axios.create({
-  baseURL: BASE_URL,
-  method: 'get',
-  headers: { 'X-Custom-Header': 'foobar' },
-});
+import instance from './axios';
 
 export const getAllPokemons = async () => {
   try {
